pr-scala-integrate-partest now actually uses the distrib being tested
Apparently, it's not enough to set scalaVersion to point to the synthetic
version of the distrib that's just been built and published to maven.

It looks that regardless of the value of scalaVersion that's being set,
sbt is going to use the compiler provided in build.sbt of scala-partest.

I was quite surprised to see that, but I made sure that it's the stale
compiler that causes the problem. I put a random println in the new
compiler, build it, published it and ran a local equivalent of
pr-scala-integrate-partest. Unfortunately, I never saw the funny println
I introduced.

Therefore I propose to switch gears and use scalaHome instead of scalaVersion.