fix links to source in scaladoc for nightly builds
The ant build assumes that providing `maven.version.suffix` means we are
building a release, and ends up using the scala version in
link-to-source urls. For nightly builds this means in the library
scaladoc we get github urls that are wrong, like
https://github.com/scala/scala/tree/v2.11.2-73fb460-nightly/src/library/scala/Predef.scala#L1.

See http://www.scala-lang.org/files/archive/nightly/2.11.x/api/2.11.2-73fb460-nightly/#scala.Predef$
for an example.

This just makes the script pass to ant the correct value: either the
scala tag of the release, or the full commit of the current HEAD.